frames2ocr/requirements.txt
pytesseract
boto3
Pillow
numpy
